Move main to app_main.
